refactor: core python package is created for allocating boostrap, config and context files (#40)
