+++ steps/clinvar_setup.py
@@ -1,6 +1,6 @@
 # steps/clinvar_setup.py
 
-from sftool.context import ExecutionContext
+from sftool.core.context import ExecutionContext
 from sftool.misc.clinvar_utils import clinvar_manager
 
 def run(ctx: ExecutionContext) -> None:
