+++ steps/sample_preprocessing.py
@@ -1,6 +1,6 @@
 # steps/sample_preprocessing_setup.py
 
-from sftool.context import ExecutionContext
+from sftool.core.context import ExecutionContext
 from sftool.misc.vcf_utils import normalize_vcf, intersect_vcf_with_bed
 from sftool.misc.pharmcat_utils import pharmCAT_vcf_preprocessor
