+++ steps/variant_selection.py
@@ -1,4 +1,4 @@
-from sftool.context import ExecutionContext
+from sftool.core.context import ExecutionContext
 from sftool.variant_selection.pr_variant_selection import pr_variant_selection
 from sftool.variant_selection.rr_variant_selection import rr_variant_selection, rr_str_selection, rr_smn1_copy_selection
